Extract TRPG.Data project for EF Core models, DbContext, and migrations
Splits persistence concerns (models, DbContext, migrations) out of the
TRPG API host into a standalone project so the data layer has no
dependency on ASP.NET Core. Also relaxes TreatWarningsAsErrors to
surface Roslyn analyzer issues as warnings instead of build failures,
and pins EF Core/Relational to 10.0.9 in TRPG.Data explicitly so that
version floor flows to consuming projects instead of the lower one
pulled in transitively by Npgsql.EntityFrameworkCore.PostgreSQL.

+++ TRPG.Tests/GraphsTests.cs
@@ -1,5 +1,5 @@
 using TRPG.Algorithms;
-using TRPG.Models;
+using TRPG.Data.Models;
 
 namespace TRPG.Tests;
 
@@ -90,7 +90,7 @@ public class GraphsTests
         // 0 S X E
         // 1 . . .
         // 2 . . .
-        var blocked = new HashSet<Point> { new(1, 0) };
+        var blocked = new HashSet<Point> { new((double)1, (double)0) };
 
         // Act
         var result = Graphs.ShortestPath(
